Move filter transforms to its own module. Convert some unittest code to pytest. Deprecate FrequencyMask.
