+++ audiomentations/__init__.py
@@ -1,3 +1,13 @@
+from .augmentations.filters import (
+    PeakingFilter,
+    LowPassFilter,
+    HighPassFilter,
+    LowShelfFilter,
+    HighShelfFilter,
+    FrequencyMask,
+    BandPassFilter,
+    BandStopFilter,
+)
 from .augmentations.spectrogram_transforms import SpecFrequencyMask, SpecChannelShuffle
 from .augmentations.transforms import (
     AddBackgroundNoise,
@@ -6,20 +16,12 @@ from .augmentations.transforms import (
     AddImpulseResponse,
     ApplyImpulseResponse,
     AddShortNoises,
-    BandPassFilter,
-    BandStopFilter,
     Clip,
     ClippingDistortion,
-    FrequencyMask,
     Gain,
-    HighPassFilter,
-    LowShelfFilter,
-    HighShelfFilter,
     LoudnessNormalization,
-    LowPassFilter,
     Mp3Compression,
     Normalize,
-    PeakingFilter,
     PitchShift,
     PolarityInversion,
     Resample,
